pyproject.toml
[project]
name = "vision-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fal-client>=0.5.9",
    "google-cloud-aiplatform>=1.36.0",
    "google-genai>=1.5.0",
    "pydotenv>=0.0.7",
    "python-dotenv>=1.0.0",
    "streamlit>=1.42.2",
]
